+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Added AppRoleAssigment related commands for service principal. [#18412]
 * Added `-WithSource` parameter to `Publish-AzBicepModule` for publishing source with a module (currently experimental)
 * Supported nullable Bicep parameters in Deployment cmdlets